+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [3.1.3] - 2018-12-17
 ### Added
+- Ability to exit Firefox with remote back button
 
 ### Changed