+++ CHANGELOG.md
@@ -4,11 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [3.1.3] - 2018-12-17
+## [3.1.3] - 2018-12-20
 ### Added
 - Ability to exit Firefox with remote back button
-
-### Changed
 
 ### Fixed
 - Issue where users were unable to exit YouTube with remote back button (#1542)